+++ src/polymer-build.ts
@@ -1,9 +1,11 @@
 export {SWConfig} from 'sw-precache';
 export {BuildAnalyzer, DepsIndex} from './analyzer';
+export {resolveBareSpecifiers} from './babel-plugin-bare-specifiers';
 export {BuildBundler} from './bundle';
 export {addCustomElementsEs5Adapter} from './custom-elements-es5-adapter';
 export {forkStream} from './fork-stream';
 export {HtmlSplitter} from './html-splitter';
+export {getOptimizeStreams, OptimizeOptions} from './optimize-streams';
 export {PolymerProject} from './polymer-project';
 export {addServiceWorker, generateServiceWorker} from './service-worker';
 export {AddServiceWorkerOptions} from './service-worker';